Fixed: add to card

+++ server/manage/root/models/root.py
@@ -30,5 +30,5 @@ class Social_Media(Abstract_Model):
 class Delivery(Abstract_Model):
 
     country = models.CharField(max_length=20, unique=True)
-    price_eur = models.IntegerField()
-    price_pln = models.IntegerField()
+    price_eur = models.FloatField()
+    price_pln = models.FloatField()
